# Smaller hashtables + block compression API