--- a/internal/lz4block/block.go
+++ b/internal/lz4block/block.go
@@ -25,23 +25,6 @@ const (
 	mfLimit = 10 + minMatch // The last match cannot start within the last 14 bytes.
 )
 
-// Pool of hash tables for CompressBlock.
-var HashTablePool = hashTablePool{sync.Pool{New: func() interface{} { return new([htSize]int) }}}
-
-type hashTablePool struct {
-	sync.Pool
-}
-
-func (p *hashTablePool) Get() *[htSize]int {
-	return p.Pool.Get().(*[htSize]int)
-}
-
-// Zero out the table to avoid non-deterministic outputs (see issue#65).
-func (p *hashTablePool) Put(t *[htSize]int) {
-	*t = [htSize]int{}
-	p.Pool.Put(t)
-}
-
 func recoverBlock(e *error) {
 	if r := recover(); r != nil && *e == nil {
 		*e = lz4errors.ErrInvalidSourceShortBuffer
@@ -68,7 +51,55 @@ func UncompressBlock(src, dst []byte) (int, error) {
 	return 0, lz4errors.ErrInvalidSourceShortBuffer
 }
 
-func CompressBlock(src, dst []byte, hashTable []int) (int, error) {
+type Compressor struct {
+	// Offsets are at most 64kiB, so we can store only the lower 16 bits of
+	// match positions: effectively, an offset from some 64kiB block boundary.
+	//
+	// When we retrieve such an offset, we interpret it as relative to the last
+	// block boundary si &^ 0xffff, or the one before, (si &^ 0xffff) - 0xffff,
+	// depending on which of these is inside the current window. If a table
+	// entry was generated more than 64kiB back in the input, we find out by
+	// inspecting the input stream.
+	table [htSize]uint16
+
+	needsReset bool
+}
+
+// Get returns the position of a presumptive match for the hash h.
+// The match may be a false positive due to a hash collision or an old entry.
+// If si < winSize, the return value may be negative.
+func (c *Compressor) get(h uint32, si int) int {
+	h &= htSize - 1
+	i := int(c.table[h])
+	i += si &^ winMask
+	if i >= si {
+		// Try previous 64kiB block (negative when in first block).
+		i -= winSize
+	}
+	return i
+}
+
+func (c *Compressor) put(h uint32, si int) {
+	h &= htSize - 1
+	c.table[h] = uint16(si)
+}
+
+var compressorPool = sync.Pool{New: func() interface{} { return new(Compressor) }}
+
+func CompressBlock(src, dst []byte) (int, error) {
+	c := compressorPool.Get().(*Compressor)
+	n, err := c.CompressBlock(src, dst)
+	compressorPool.Put(c)
+	return n, err
+}
+
+func (c *Compressor) CompressBlock(src, dst []byte) (int, error) {
+	if c.needsReset {
+		// Zero out reused table to avoid non-deterministic output (issue #65).
+		c.table = [htSize]uint16{}
+	}
+	c.needsReset = true // Only false on first call.
+
 	// Return 0, nil only if the destination buffer size is < CompressBlockBound.
 	isNotCompressible := len(dst) < CompressBlockBound(len(src))
 
@@ -85,15 +116,6 @@ func CompressBlock(src, dst []byte, hashTable []int) (int, error) {
 		goto lastLiterals
 	}
 
-	if cap(hashTable) < htSize {
-		poolTable := HashTablePool.Get()
-		defer HashTablePool.Put(poolTable)
-		hashTable = poolTable[:]
-	} else {
-		hashTable = hashTable[:htSize]
-	}
-	_ = hashTable[htSize-1]
-
 	// Fast scan strategy: the hash table only stores the last 4 bytes sequences.
 	for si < sn {
 		// Hash the next 6 bytes (sequence)...
@@ -103,33 +125,30 @@ func CompressBlock(src, dst []byte, hashTable []int) (int, error) {
 
 		// We check a match at s, s+1 and s+2 and pick the first one we get.
 		// Checking 3 only requires us to load the source one.
-		ref := hashTable[h]
-		ref2 := hashTable[h2]
-		hashTable[h] = si
-		hashTable[h2] = si + 1
+		ref := c.get(h, si)
+		ref2 := c.get(h2, si)
+		c.put(h, si)
+		c.put(h2, si+1)
+
 		offset := si - ref
 
-		// If offset <= 0 we got an old entry in the hash table.
-		if offset <= 0 || offset >= winSize || // Out of window.
-			uint32(match) != binary.LittleEndian.Uint32(src[ref:]) { // Hash collision on different matches.
+		if ref < 0 || uint32(match) != binary.LittleEndian.Uint32(src[ref:]) {
 			// No match. Start calculating another hash.
 			// The processor can usually do this out-of-order.
 			h = blockHash(match >> 16)
-			ref = hashTable[h]
+			ref3 := c.get(h, si+2)
 
 			// Check the second match at si+1
 			si += 1
 			offset = si - ref2
 
-			if offset <= 0 || offset >= winSize ||
-				uint32(match>>8) != binary.LittleEndian.Uint32(src[ref2:]) {
+			if ref2 < 0 || uint32(match>>8) != binary.LittleEndian.Uint32(src[ref2:]) {
 				// No match. Check the third match at si+2
 				si += 1
-				offset = si - ref
-				hashTable[h] = si
+				offset = si - ref3
+				c.put(h, si)
 
-				if offset <= 0 || offset >= winSize ||
-					uint32(match>>16) != binary.LittleEndian.Uint32(src[ref:]) {
+				if ref3 < 0 || uint32(match>>16) != binary.LittleEndian.Uint32(src[ref3:]) {
 					// Skip one extra byte (at si+3) before we check 3 matches again.
 					si += 2 + (si-anchor)>>adaptSkipLog
 					continue
@@ -221,7 +240,7 @@ func CompressBlock(src, dst []byte, hashTable []int) (int, error) {
 		}
 		// Hash match end-2
 		h = blockHash(binary.LittleEndian.Uint64(src[si-2:]))
-		hashTable[h] = si - 2
+		c.put(h, si-2)
 	}
 
 lastLiterals:
@@ -269,7 +288,30 @@ func blockHashHC(x uint32) uint32 {
 	return x * hasher >> (32 - winSizeLog)
 }
 
-func CompressBlockHC(src, dst []byte, depth CompressionLevel, hashTable, chainTable []int) (_ int, err error) {
+type CompressorHC struct {
+	// hashTable: stores the last position found for a given hash
+	// chainTable: stores previous positions for a given hash
+	hashTable, chainTable [htSize]int
+	needsReset            bool
+}
+
+var compressorHCPool = sync.Pool{New: func() interface{} { return new(CompressorHC) }}
+
+func CompressBlockHC(src, dst []byte, depth CompressionLevel) (int, error) {
+	c := compressorHCPool.Get().(*CompressorHC)
+	n, err := c.CompressBlock(src, dst, depth)
+	compressorHCPool.Put(c)
+	return n, err
+}
+
+func (c *CompressorHC) CompressBlock(src, dst []byte, depth CompressionLevel) (_ int, err error) {
+	if c.needsReset {
+		// Zero out reused table to avoid non-deterministic output (issue #65).
+		c.hashTable = [htSize]int{}
+		c.chainTable = [htSize]int{}
+	}
+	c.needsReset = true // Only false on first call.
+
 	defer recoverBlock(&err)
 
 	// Return 0, nil only if the destination buffer size is < CompressBlockBound.
@@ -286,25 +328,6 @@ func CompressBlockHC(src, dst []byte, depth CompressionLevel, hashTable, chainTa
 		goto lastLiterals
 	}
 
-	// hashTable: stores the last position found for a given hash
-	// chainTable: stores previous positions for a given hash
-	if cap(hashTable) < htSize {
-		poolTable := HashTablePool.Get()
-		defer HashTablePool.Put(poolTable)
-		hashTable = poolTable[:]
-	} else {
-		hashTable = hashTable[:htSize]
-	}
-	_ = hashTable[htSize-1]
-	if cap(chainTable) < htSize {
-		poolTable := HashTablePool.Get()
-		defer HashTablePool.Put(poolTable)
-		chainTable = poolTable[:]
-	} else {
-		chainTable = chainTable[:htSize]
-	}
-	_ = chainTable[htSize-1]
-
 	if depth == 0 {
 		depth = winSize
 	}
@@ -317,7 +340,7 @@ func CompressBlockHC(src, dst []byte, depth CompressionLevel, hashTable, chainTa
 		// Follow the chain until out of window and give the longest match.
 		mLen := 0
 		offset := 0
-		for next, try := hashTable[h], depth; try > 0 && next > 0 && si-next < winSize; next, try = chainTable[next&winMask], try-1 {
+		for next, try := c.hashTable[h], depth; try > 0 && next > 0 && si-next < winSize; next, try = c.chainTable[next&winMask], try-1 {
 			// The first (mLen==0) or next byte (mLen>=minMatch) at current match length
 			// must match to improve on the match length.
 			if src[next+mLen] != src[si+mLen] {
@@ -344,8 +367,8 @@ func CompressBlockHC(src, dst []byte, depth CompressionLevel, hashTable, chainTa
 			offset = si - next
 			// Try another previous position with the same hash.
 		}
-		chainTable[si&winMask] = hashTable[h]
-		hashTable[h] = si
+		c.chainTable[si&winMask] = c.hashTable[h]
+		c.hashTable[h] = si
 
 		// No match found.
 		if mLen == 0 {
@@ -364,8 +387,8 @@ func CompressBlockHC(src, dst []byte, depth CompressionLevel, hashTable, chainTa
 			match >>= 8
 			match |= uint32(src[si+3]) << 24
 			h := blockHashHC(match)
-			chainTable[si&winMask] = hashTable[h]
-			hashTable[h] = si
+			c.chainTable[si&winMask] = c.hashTable[h]
+			c.hashTable[h] = si
 			si++
 		}