This is the PR promised in #92: uint64-typed hash tables for block compression (except in HC mode). It introduces a new API for block compression, while marking the old one deprecated, while also fixing some issues with the benchmarks. See the commit messages for details.

Benchmark results on amd64:

```
name              old time/op    new time/op    delta
Compress-8          3.13ms ± 3%    3.00ms ± 0%    -4.10%  (p=0.000 n=21+20)
CompressRandom-8    17.9µs ± 1%     8.6µs ± 1%   -51.78%  (p=0.000 n=19+20)
CompressHC-8        10.9ms ± 1%    10.2ms ± 1%    -6.12%  (p=0.000 n=21+20)
CompressPg1661-8     685µs ± 3%     699µs ± 2%    +2.05%  (p=0.000 n=21+21)
CompressDigits-8     649µs ± 3%     645µs ± 4%      ~     (p=0.083 n=19+21)
CompressTwain-8      676µs ± 2%     681µs ± 3%      ~     (p=0.078 n=20+21)
CompressRand-8       641µs ± 3%     627µs ± 2%    -2.18%  (p=0.000 n=21+21)

name              old alloc/op   new alloc/op   delta
Compress-8           923B ±100%        0B       -100.00%  (p=0.000 n=21+21)
CompressRandom-8    5.86B ±100%     0.00B       -100.00%  (p=0.000 n=21+21)
CompressHC-8        4.80kB ± 1%    0.00kB       -100.00%  (p=0.000 n=16+21)
CompressPg1661-8    4.25MB ± 1%    4.27MB ± 1%    +0.46%  (p=0.000 n=21+21)
CompressDigits-8    4.27MB ± 1%    4.25MB ± 1%    -0.58%  (p=0.000 n=21+20)
CompressTwain-8     4.25MB ± 1%    4.27MB ± 1%    +0.48%  (p=0.000 n=21+21)
CompressRand-8      4.26MB ± 0%    4.24MB ± 1%    -0.51%  (p=0.000 n=19+21)

name              old allocs/op  new allocs/op  delta
Compress-8            0.00           0.00           ~     (all equal)
CompressRandom-8      0.00           0.00           ~     (all equal)
CompressHC-8          0.00           0.00           ~     (all equal)
CompressPg1661-8      4.00 ± 0%      4.00 ± 0%      ~     (all equal)
CompressDigits-8      4.00 ± 0%      4.00 ± 0%      ~     (all equal)
CompressTwain-8       4.00 ± 0%      4.00 ± 0%      ~     (all equal)
CompressRand-8        4.00 ± 0%      4.00 ± 0%      ~     (all equal)

name              old speed      new speed      delta
CompressRandom-8   917MB/s ± 1%  1902MB/s ± 1%  +107.39%  (p=0.000 n=19+20)
CompressPg1661-8   868MB/s ± 3%   851MB/s ± 2%    -2.01%  (p=0.000 n=21+21)
CompressDigits-8   154MB/s ± 3%   155MB/s ± 4%      ~     (p=0.051 n=20+21)
CompressTwain-8    573MB/s ± 4%   570MB/s ± 3%      ~     (p=0.153 n=21+21)
CompressRand-8    25.6MB/s ± 3%  26.1MB/s ± 2%    +2.21%  (p=0.000 n=21+21)

name              old outbytes   new outbytes   delta
Compress-8            329k ± 0%      329k ± 0%      ~     (p=1.000 n=1+1)
CompressRandom-8      0.00           0.00           ~     (all equal)
CompressHC-8          288k ± 0%      288k ± 0%      ~     (all equal)
CompressPg1661-8      329k ± 0%      329k ± 0%      ~     (p=1.000 n=1+1)
CompressDigits-8      100k ± 0%      100k ± 0%      ~     (all equal)
CompressTwain-8       227k ± 0%      227k ± 0%      ~     (p=1.000 n=1+1)
CompressRand-8       16.4k ± 0%     16.4k ± 0%      ~     (all equal)
```

On the Raspberry Pi 4:

```
name              old time/op    new time/op    delta
Compress-4          14.9ms ± 2%    16.8ms ± 2%   +12.45%  (p=0.000 n=20+20)
CompressRandom-4     126µs ± 2%     126µs ± 2%      ~     (p=0.758 n=20+20)
CompressHC-4        44.7ms ± 4%    49.0ms ± 2%    +9.74%  (p=0.000 n=19+17)
CompressPg1661-4    2.78ms ±29%    2.71ms ± 3%      ~     (p=0.178 n=16+20)
CompressDigits-4    1.90ms ± 4%    1.89ms ± 2%      ~     (p=0.631 n=16+17)
CompressTwain-4     2.33ms ± 3%    2.33ms ± 3%      ~     (p=0.931 n=19+19)
CompressRand-4      1.88ms ± 4%    1.89ms ± 2%      ~     (p=0.196 n=20+18)

name              old alloc/op   new alloc/op   delta
Compress-4         1.86kB ±100%    0.00kB       -100.00%  (p=0.000 n=20+19)
CompressRandom-4    19.5B ±100%      0.0B       -100.00%  (p=0.000 n=20+20)
CompressHC-4       5.59kB ±100%    0.00kB       -100.00%  (p=0.000 n=20+20)
CompressPg1661-4    4.20MB ± 0%    4.20MB ± 1%      ~     (p=0.429 n=20+20)
CompressDigits-4    4.20MB ± 0%    4.20MB ± 0%      ~     (p=0.255 n=19+19)
CompressTwain-4     4.20MB ± 0%    4.19MB ± 0%    -0.12%  (p=0.036 n=20+19)
CompressRand-4      4.20MB ± 0%    4.20MB ± 0%      ~     (p=0.659 n=20+20)

name              old allocs/op  new allocs/op  delta
Compress-4            0.00           0.00           ~     (all equal)
CompressRandom-4      0.00           0.00           ~     (all equal)
CompressHC-4          0.00           0.00           ~     (all equal)
CompressPg1661-4      4.00 ± 0%      4.00 ± 0%      ~     (all equal)
CompressDigits-4      4.00 ± 0%      4.00 ± 0%      ~     (all equal)
CompressTwain-4       4.00 ± 0%      4.00 ± 0%      ~     (all equal)
CompressRand-4        4.00 ± 0%      4.00 ± 0%      ~     (all equal)

name              old speed      new speed      delta
CompressRandom-4   130MB/s ± 2%   130MB/s ± 2%      ~     (p=0.753 n=20+20)
CompressPg1661-4   215MB/s ±23%   220MB/s ± 3%      ~     (p=0.178 n=16+20)
CompressDigits-4  52.7MB/s ± 4%  52.8MB/s ± 2%      ~     (p=0.612 n=16+17)
CompressTwain-4    167MB/s ± 3%   166MB/s ± 4%      ~     (p=0.713 n=19+20)
CompressRand-4    8.71MB/s ± 4%  8.67MB/s ± 3%      ~     (p=0.214 n=20+18)
```